requirements.txt
dbt-core==1.8.3
dbt-snowflake==1.8.3
dbt-duckdb==1.8.2
sqlmesh==0.110.0
sqlmesh[snowflake]
sqlmesh[web]
sqlmesh[github]
pandas==2.2.2
faker==26.0.0
python-dotenv==1.0.1
pyarrow==16.1.0
progressbar==2.5
pyyaml==6.0.1